Optimization 2/2

+++ TP6/TP6_Project/Project/pid_regulator.h
@@ -7,7 +7,7 @@ void pid_regulator_start(void);
 
 float speedcms_to_speedsteps (uint8_t speed_cms);
 
-void motor_set_position(float position_r, float position_l, float speed_r, float speed_l);
+void motor_set_position(float position_r, float position_l, int16_t speed_r, int16_t speed_l);
 
 void mov_circ_right(float vitesse,float rayon,float angle, int mode);
